Fix compilation of iconv tests with libcxx 19
[libcxx 18 deprecated the `std::basic_string` template specializations
for `int16_t` and `int32_t`][1], which are currently used by the iconv
tests. [The template specializations were removed altogether in libcxx
19][2], which breaks compilation of the iconv tests. Replace `int16_t`
and `int32_t` with `char16_t` and `char32_t` to fix compilation with
libcxx 19.

I added the `static_assert` for checking the sizes of `char16_t` and
`char32_t` as [the standard allows them to contain padding bytes][3].

[1]: https://reviews.llvm.org/D138307
[2]: https://reviews.llvm.org/D157058
[3]: https://en.cppreference.com/w/cpp/language/types#char16_t